Separate redraw from refresh.
Previously, system would only refresh when the screen was awake.
This caused problems for the network rate calculations that expect
the refresh interval to be somewhat consistent,
it would sum all of the activity since the last time the screen was on.

Now, all stats are refreshed regularly, but the screen is still
only redrawn when "awake".